implements the inbox with CRUD and sqlalchemy
This implements an inbox, with support in the streamlit application, for CRUD operations on investigation reports. It also changes the investigation to automatically save the report into the database.

In the future there will need to be support for not waiting for the investigation to complete (it might take longer than the HTTP timeout). It'll also need to add support for other sqlalchemy env vars for connections to DBs with user/pw/etc.